석현지 커밋앤푸시 5:41PM

+++ WebContent/WEB-INF/view/customer/QA/QADetail.jsp
@@ -1,3 +1,4 @@
+<%@page import="sports.com.util.AES256Util"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%@ page import="sports.com.util.CmmUtil" %>
@@ -12,22 +13,21 @@ if (rDTO==null) {
 }
 
 String qa_no = CmmUtil.nvl(request.getParameter("qa_no"));
-String ss_user_no = CmmUtil.nvl((String)session.getAttribute("SESSION_USER_NO"));
+String user_no = CmmUtil.nvl((String)session.getAttribute("user_no"));
 
 int edit = 1;
 
-if (ss_user_no.equals("")) {
+if (user_no.equals("")) {
 	
 	edit = 3;
 	
-} else if (ss_user_no.equals(CmmUtil.nvl(rDTO.getReg_user_no()))) {
+} else if (user_no.equals(CmmUtil.nvl(rDTO.getReg_user_no()))) {
 	
 	edit = 2;
 	
 }
 
 System.out.println("user_no: " + CmmUtil.nvl(rDTO.getReg_user_no()));
-System.out.println("ss_user_no: " + ss_user_no);
 %>
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml">
@@ -122,7 +122,7 @@ function doList() {
 		</tr>
 	
 		<tr>
-			<td align="left" colspan="2">작성자&nbsp;&nbsp;<%=CmmUtil.nvl(rDTO.getUser_name())%>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;작성일&nbsp;&nbsp;<%=CmmUtil.nvl(rDTO.getReg_dt())%></td>
+			<td align="left" colspan="2">작성자&nbsp;&nbsp;<%=AES256Util.strDecode(CmmUtil.nvl(rDTO.getUser_name()))%>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;작성일&nbsp;&nbsp;<%=CmmUtil.nvl(rDTO.getReg_dt())%></td>
 		</tr>	
 	
 		<tr>
@@ -140,10 +140,15 @@ function doList() {
 	
 			</div>
 		</div>
+		
+	<% if (user_no.equals(CmmUtil.nvl(rDTO.getReg_user_no()))) {%>
 	
-	<input type="button" value="수정" onclick="location.href='javascript:doEdit();' " />
-	<input type="button" value="삭제" onclick="location.href='javascript:doDelete();' " />
-	<input type="button" value="목록" onclick="location.href='javascript:doList();' " />
+		<input type="button" value="수정" onclick="location.href='javascript:doEdit();' " />
+		<input type="button" value="삭제" onclick="location.href='javascript:doDelete();' " />
+		
+	<%} %>
+	
+		<input type="button" value="목록" onclick="location.href='javascript:doList();' " />
 
 </form>
